Added a lot of extra features and made the program look a lot nicer than before
features:
now correctly shows the amout of bits are down- and uploaded.
Textboxes can scroll now
Added way to allow 12k of data being pushed every second to the textboxes without the program crashing

TODO:
Add a way to dynamically resize the textbox textarrays so that it only grows when it needs to (improve memory management).